requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
python-dateutil==2.9.0.post0
tqdm==4.66.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2